Refactor RoboteqCommander into a base class of Roboteq Command Generator
We need a class that is agnostic to using the StringIO Object because we 
will want to output commands to serial port objects and network ports, 
and objects with the like of different protocols. All methods can easily 
be overwritten. General RoboteQ comamnder class can still operate on 
Python StringIO object, which is currently what unit testing suite uses. 
This is good, functionality can be added to command generator in order 
to support things like maintance commands, or when a proper dictionary 
structure that checks for argument correctness is implemented. Error 
checking and exception handling also still needs to be implemented.