change behavior of roboteq command
I am trying to best redesign roboteq command for best use practices. I 
am trying to consider the best way to do this so that it can be adapted 
for different mediums for output. 

For right now I am considering some kind of paradigm where it has output 
formatting as the final step, but that leads to a lot of repeated code. 
It should have a funciton that is uniform for checking code. But I would 
really hate to have a formatting and another funciton. Perhaps it would 
be best implemented as a helper funciton?