Implement construct, format, submit paradigm
The roboteq commander now uses a new paradigm that has you construct, 
format, and submit commands to the device. However, how exactly this 
should be implemented is still up in the air. The next step should be 
to, instead of changing it directly in the definition, to just write the 
unittests for the implementation and then make the implementaiton fit 
that and be done with it.