Add RoboteQ Device Class skeleton
Yes, I know it is ugly and bad, however I am planning on intentionally 
doing it this way in order to accurately represent how our controllers 
break down in the RoboteQ design process, and impose the same order on 
them. I will go back in and write proper class methods, so that they can 
all be accesed using just one class and all of the internal structure is 
hidden from users of the module.